# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Interview Appointment API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/slots": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "List all slots",
                "operationId": "52b2e5f0f5a47bdeebd33aee49091640",
                "parameters": [
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Start time filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "End time filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "time"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Pagination limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of slots",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "Recode": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Slot"
                                                    }
                                                },
                                                "Pagination_Limit": {
                                                    "type": "integer"
                                                },
                                                "Current_Page": {
                                                    "type": "integer"
                                                },
                                                "Total_Recode": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No slots found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Slots"
                ],
                "summary": "Create a new slot",
                "operationId": "072807e7de7dcdb5aef91c18304bcb3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SlotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slot created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to create slot"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/slots/{id}": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "Display a specific slot",
                "operationId": "f40ae8b7c5cf60f2aa364314df03f567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Slot ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of slot",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Slot"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Slots"
                ],
                "summary": "Update a specific slot",
                "operationId": "21bc050040565f08041471cfd17fd7ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Slot ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SlotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slot updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Slots"
                ],
                "summary": "Delete a specific slot",
                "operationId": "85d7be2ceaa0aba0893f3b2a7f59ff49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Slot ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slot not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SlotRequest": {
                "required": [
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "17:00:00"
                    }
                },
                "type": "object"
            },
            "Slot": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token to access protected routes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}